Add files via upload
'League of Legends' is a game of the genre MOBA that came out in October 2009. It has been almost ten years since the flagship of the developer Riot Games was launched, and it is still a success. This tool allows you to perform ping requests to LoL server endpoints. You can use this tool to measure your LoL ping before entering the game.